pyproject.toml
[project]
name = "debateclub"
version = "0.1.0"
description = "Two LLMs debate a topic, a judge scores them"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[project.scripts]
debateclub = "debateclub.cli:main"

[tool.setuptools.packages.find]
include = ["debateclub*"]

[tool.setuptools.package-data]
debateclub = ["static/**/*", "templates/**/*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
